Remove useless files in gen package and add a method in Generator to generate a sequence Diagram and implement it in PlantUmlGenerator.
