+++ src/main/java/com/github/forax/umldoc/editor/MermaidCmdLineParser.java
@@ -1,6 +1,6 @@
 package com.github.forax.umldoc.editor;
 
-import com.github.forax.umldoc.gen.ClassDiagramMermaidGenerator;
+import com.github.forax.umldoc.gen.MermaidGenerator;
 import java.util.Optional;
 import picocli.CommandLine;
 import picocli.CommandLine.Option;
@@ -43,7 +43,7 @@ public class MermaidCmdLineParser implements CommandLineParser {
       var options = new OptionsCmd();
       var args = line.substring("%% umldoc".length() + 1).split(" ");
       new CommandLine(options).execute(args);
-      GeneratorConfiguration.filterPackage(options.packageName, new ClassDiagramMermaidGenerator());
+      GeneratorConfiguration.filterPackage(options.packageName, new MermaidGenerator());
       return Optional.empty();
     }
