+++ src/main/java/com/github/forax/umldoc/gen/Generator.java
@@ -2,6 +2,7 @@ package com.github.forax.umldoc.gen;
 
 import com.github.forax.umldoc.core.Dependency;
 import com.github.forax.umldoc.core.Entity;
+import com.github.forax.umldoc.core.Method;
 import java.io.IOException;
 import java.io.Writer;
 import java.util.List;
@@ -22,21 +23,18 @@ public interface Generator {
   void generate(boolean header, List<Entity> entities, List<Dependency> dependencies,
                             Writer writer) throws IOException;
 
-  static void addHeader(boolean header, Writer writer) throws IOException {
-    if (header) {
-      writer.append("""
-          @startuml
-          
-          """);
-    }
-  }
-
-  static void addFooter(boolean header, Writer writer) throws IOException {
-    if (header) {
-      writer.append("""
-          
-          @enduml
-          """);
-    }
-  }
+  /**
+   * Generate a Sequence diagram from an entry point method.
+   * This method must be configured by the user.
+   * Eg : Could be the main() method.
+   *
+   * @param header if a header/footer should be generated
+   * @param entryEntity The entity where the entry point method is located
+   * @param entryPoint the first method that will appear in the sequence diagram.
+   *                   This method must be configured by the user.
+   * @param writer the writer
+   * @throws IOException if an I/O exception occurs
+   */
+  void generateSequenceDiagram(boolean header, Entity entryEntity, Method entryPoint,
+                               Writer writer) throws IOException;
 }
