+++ src/main/java/com/github/forax/umldoc/gen/MermaidGenerator.java
@@ -6,15 +6,30 @@ import com.github.forax.umldoc.core.AssociationDependency;
 import com.github.forax.umldoc.core.AssociationDependency.Cardinality;
 import com.github.forax.umldoc.core.Dependency;
 import com.github.forax.umldoc.core.Entity;
+import com.github.forax.umldoc.core.Entity.Stereotype;
 import com.github.forax.umldoc.core.Field;
+import com.github.forax.umldoc.core.Method;
+import com.github.forax.umldoc.core.TypeInfo;
 import java.io.IOException;
 import java.io.Writer;
 import java.util.List;
+import java.util.Locale;
 
 /**
- * Generate a class diagram using the plantuml format.
+ * Generate a class diagram using the mermaid format.
  */
-public final class ClassDiagramPlantUmlGenerator implements Generator {
+public final class MermaidGenerator implements Generator {
+
+  private static String shortName(String name) {
+    var index = name.lastIndexOf('.');
+    var localName = index == -1 ? name : name.substring(index + 1);
+    return localName.replaceAll("[^\\p{Alnum}]", "_");
+  }
+
+  private static String shortName(TypeInfo typeInfo) {
+    return shortName(typeInfo.name());
+  }
+
   private static String cardinality(Cardinality cardinality) {
     return switch (cardinality) {
       case ONLY_ONE -> "1";
@@ -27,13 +42,12 @@ public final class ClassDiagramPlantUmlGenerator implements Generator {
       throws IOException {
     for (var dependency : dependencies) {
       if (dependency instanceof AssociationDependency association) {
-        writer.append(
-            """
+        writer.append("""
             %s --> "%s" %s : %s
             """.formatted(
-                association.left().entity().type().name(),
+                shortName(association.left().entity().type()),
                 cardinality(association.right().cardinality()),
-                association.right().entity().type().name(),
+                shortName(association.right().entity().type()),
                 association.right().label().orElse("")));
         continue;
       }
@@ -45,25 +59,30 @@ public final class ClassDiagramPlantUmlGenerator implements Generator {
     }
   }
 
-  private static String entityType(Entity.Stereotype stereotype) {
+  private static String toStereotypeText(Stereotype stereotype) {
     return switch (stereotype) {
-      case ANNOTATION, INTERFACE -> "interface";
-      case ABSTRACT -> "abstract";
-      case CLASS, RECORD -> "class";
-      case ENUM -> "enum";
+      case ABSTRACT, ANNOTATION, INTERFACE, RECORD ->
+          "<<" + stereotype.name().toLowerCase(Locale.ROOT) + ">>";
+      case ENUM -> "<<enumeration>>";
+      case CLASS -> throw new AssertionError();
     };
   }
 
   private static void generateEntities(List<Entity> entities, Writer writer)
       throws IOException {
     for (var entity : entities) {
-      var stereotype = entity.stereotype();
-      var entityType = entityType(stereotype);
       writer.append("""
-          %s %s {
-          """.formatted(entityType, entity.type()));
+          class %s {
+          """.formatted(shortName(entity.type())));
 
-      if (stereotype == Entity.Stereotype.ENUM) {
+      var stereotype = entity.stereotype();
+      if (stereotype != Stereotype.CLASS) {
+        writer.append("""
+              %s
+            """.formatted(toStereotypeText(stereotype)));
+      }
+
+      if (entity.stereotype() == Stereotype.ENUM) {
         generateEnumConstants(entity.fields(), writer);
       } else {
         generateFields(entity.fields(), writer);
@@ -96,15 +115,25 @@ public final class ClassDiagramPlantUmlGenerator implements Generator {
 
   @Override
   public void generate(boolean header, List<Entity> entities, List<Dependency> dependencies,
-                                   Writer writer) throws IOException {
+                       Writer writer) throws IOException {
     requireNonNull(entities);
     requireNonNull(dependencies);
     requireNonNull(writer);
-    Generator.addHeader(header, writer);
+    if (header) {
+      writer.append("""
+          classDiagram
+          direction TB
+          
+          """);
+    }
 
     generateEntities(entities, writer);
     generateDependencies(dependencies, writer);
+  }
 
-    Generator.addFooter(header, writer);
+  @Override
+  public void generateSequenceDiagram(boolean header, Entity entryEntity, Method entryPoint,
+                                      Writer writer) throws IOException {
+    throw new UnsupportedOperationException("TODO");
   }
 }
