--- src/main/java/com/github/forax/umldoc/gen/SequenceDiagramMermaidGenerator.java
@@ -1,15 +1,0 @@
-package com.github.forax.umldoc.gen;
-
-import com.github.forax.umldoc.core.Dependency;
-import com.github.forax.umldoc.core.Entity;
-import java.io.IOException;
-import java.io.Writer;
-import java.util.List;
-
-public class SequenceDiagramMermaidGenerator implements Generator {
-  @Override
-  public void generate(boolean header, List<Entity> entities, List<Dependency> dependencies,
-                       Writer writer) throws IOException {
-    // TODO
-  }
-}
